+++ src/use-async-combine-seq.js
@@ -27,7 +27,8 @@ export const useAsyncCombineSeq = (...asyncTasks) => {
   return {
     ...task,
     pending: asyncTasks.some(({ pending }) => pending),
-    error: asyncTasks.map(({ error }) => error),
+    error: asyncTasks.find(({ error }) => error),
+    errorAll: asyncTasks.map(({ error }) => error),
     result: asyncTasks.map(({ result }) => result),
   };
 };