+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Changed
+- Fix combine return type
 
 ## [0.2.0] - 2018-12-14
 ### Added